refactor(web): split api integration suites [OPE-333]
