+++ crates/opengoose-web/src/tests/api_integration/mod.rs
@@ -1,0 +1,3 @@
+mod alerts_routes;
+mod core_routes;
+mod support;
